pyproject.toml
[project]
name = "minutes_iq"
version = "0.1.0"
description = "MinutesIQ is a role-based backend platform that ingests public meeting records, applies configurable keyword intelligence, and delivers structured reports to authenticated users."
requires-python = ">=3.12"

dependencies = [
    "numpy>=0",
    "pandas>=0",
    "requests>=0",
    "beautifulsoup4>=0",
    "lxml>=0",
    "nltk>=0",
    "tqdm>=0",
    "spacy>=0",
    "catalogue>=0",
    "cymem>=0",
    "preshed>=0",
    "thinc>=0",
    "cryptography>=0",
    "pdfminer-six>=0",
    "pdfplumber>=0",
    "pillow>=0",
    "pypdfium2>=0",
    "fastapi>=0.123.10",
    "httpx>=0.28.1",
    "pandas-stubs==2.3.3.251201",
    "streamlit>=1.52.1",
    "fitz>=0.0.1.dev2",
    "bcrypt>=3.2.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.38.0",
    "libsql-client>=0.3.1",
    "libsql-experimental>=0.0.55",
    "mypy>=1.19.1",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "pytest>=7.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88
fix = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "W",  # pycodestyle warnings
    "I",  # import sorting (replaces isort)
    "B",  # bugbear
    "UP", # pyupgrade
]
ignore = [
    "E203", # whitespace before ':', black/ruff compatible
    "E501", # allow long lines (hashes, tests, etc.)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
# Ignore missing library stubs for third-party packages
ignore_missing_imports = true
